Fixed "jumping" bug
Changed the node position calculation to compute desired position for
all nodes before calling SetPosition on any of them. This prevents the
case where half an iteration for a given node executes before the node
acts on the new position, leading to a quick jump in position when only
some connections are handled.