Controller Focus and Focused Camera
There are now functions OnBlur and OnFocus for handling changes in
controller focus. Currently, controllers do not store whether they are
focused because Camera needs to track it separately for use with hit
testing.

On a related note, there is now a focused camera. If a camera is not
focused, it will not handle input for movement and selection.