Made Resizer General Purpose
Resizer is now a script that can be added to any GUI object to hook into
the CursorManager, updating the current image. The drag event can now be
subscribed to externally, so no dependencies are required from Resizer
to other classes. Currently Resizer still uses the old panel member to
send events, but that will be changed soon.